upload_service/pyproject.toml
[project]
name = "upload-service"
version = "0.1.0"
description = "Upload media to S3"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "boto3>=1.38.36",
    "fastapi>=0.115.12",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.34.3",
]
